Indicator name component.
